consolidate bucket edit and create modal
