+++ apps/studio/components/interfaces/Storage/index.ts
@@ -1,5 +1,6 @@
-export { default as StorageExplorer } from './StorageExplorer/StorageExplorer'
-export { default as StoragePolicies } from './StoragePolicies/StoragePolicies'
-export { default as StorageSettings } from './StorageSettings/StorageSettings'
+// Unified bucket modal components
+export { BucketModal, CreateBucketModal, EditBucketModal } from './BucketModal'
 
-export { DeleteBucketModal } from './DeleteBucketModal'
+// Legacy exports for backward compatibility (deprecated)
+export { default as CreateBucketModalLegacy } from './CreateBucketModal'
+export { EditBucketModal as EditBucketModalLegacy } from './EditBucketModal'
